requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
lxml==4.9.3
aiohttp==3.9.1
SQLAlchemy==2.0.23
APScheduler==3.10.4
fake-useragent==1.4.0
httpx==0.25.2
selenium==4.18.1 